pull limit notifications into their own component for re-use
